# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Product Manager Copilot Assist API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List product opportunities
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: source_type
          in: query
          schema:
            type: string
        - name: sentiment
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of Product Ideas to return (up to 200).
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of Product Ideas to skip for pagination.
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create a product opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get a product opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
        "404":
          description: Not found
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update a product opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete a product opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /opportunities/{id}/analyze:
    post:
      operationId: analyzeOpportunity
      tags: [opportunities]
      summary: Run AI analysis on an opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated opportunity with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"

  # --- Signals (raw inputs) ---
...
components:
...
  schemas:
...
    Opportunity:
      type: object
      required: [id, title, description, status, sourceType, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        sourceType:
          type: string
        aiSummary:
          type: ["string", "null"]
        customerProblem:
          type: ["string", "null"]
        suggestedSolution:
          type: ["string", "null"]
        businessValue:
          type: ["string", "null"]
        estimatedCustomerImpact:
          type: ["string", "null"]
        estimatedBusinessImpact:
          type: ["string", "null"]
        urgency:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        sentiment:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, under_review, ready_for_prioritization, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OpportunityDetail:
      allOf:
        - $ref: "#/components/schemas/Opportunity"
        - type: object
          properties:
            evidence:
              $ref: "#/components/schemas/EvidencePanel"
            relatedSignals:
              type: array
              items:
                $ref: "#/components/schemas/Signal"
    EvidencePanel:
      type: object
      properties:
        customerRequestCount:
          type: number
        stakeholderMentions:
          type: number
        meetingMentions:
          type: number
        competitorReferences:
          type: number
        socialMentions:
          type: number
        exampleQuotes:
          type: array
          items:
            type: string
        sourceLinks:
          type: array
          items:
            type: string
    OpportunityInput:
      type: object
      required: [title, description, sourceType]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        sourceType:
          type: string
        originalContent:
          type: string
        customerProblem:
          type: string
        suggestedSolution:
          type: string
        businessValue:
          type: string
        urgency:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, under_review, ready_for_prioritization, archived]
    OpportunityPatch:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        sourceType:
          type: string
        customerProblem:
          type: string
        suggestedSolution:
          type: string
        businessValue:
          type: string
        urgency:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, under_review, ready_for_prioritization, archived]

    # --- Signal ---
    Signal:
      type: object
      required: [id, content, sourceType, processed, createdAt]
      properties:
        id:
          type: number
        content:
          type: string
        sourceType:
          type: string
        sourcePlatform:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
        processed:
          type: boolean
        opportunityId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...